+++ src/cli/commands/info.rs
@@ -1,5 +1,11 @@
 // SPDX-License-Identifier: GPL-3.0-only
-use crate::cli::display::{standard_print::DisplayJoined, styled::MapStyled};
+use crate::{
+    cli::display::{standard_print::DisplayJoined, styled::MapStyled},
+    installer::types::PackageName,
+    platforms::{DEFAULT_CONFIG_DIR, OsVersion, Target},
+    repositories::manager::RepositoryManager,
+    utils::packit_version::{current_packit_version, packit_version_name},
+};
 use clap::Args;
 use colored::Colorize;
 use std::process::exit;
@@ -28,14 +34,14 @@ use crate::{
 #[derive(Args, Debug)]
 pub struct InfoArgs {
     /// Optional package id
-    package: OptionalPackageId,
+    package: Option<OptionalPackageId>,
 
     /// True if verbose information should be shown
     #[arg(short, long, default_value = "false")]
     verbose: bool,
 
     /// True if displaying package trees as well
-    #[arg(long, default_value = "false")]
+    #[arg(long, default_value = "false", requires = "package")]
     tree: bool,
 }
 
@@ -45,15 +51,20 @@ impl HandleCommand for InfoArgs {
         let register_dir = PackageRegister::get_path(&config.prefix_directory);
         let register = PackageRegister::from(&register_dir).unwrap_or_exit(1);
 
+        let Some(package) = &self.package else {
+            self.display_packit_info(&config, &register);
+            return;
+        };
+
         // Get package information
-        let package = match register.get_package(&self.package.name) {
+        let installed_package = match register.get_package(&package.name) {
             Some(package) => package,
-            None => not_found::register_package(&self.package.name, &register),
+            None => not_found::register_package(&package.name, &register),
         };
 
         // Display tree if tree flag is given
         if self.tree {
-            let Some(package_id) = self.package.versioned() else {
+            let Some(package_id) = package.versioned() else {
                 error!(msg: "Displaying a tree requires package version to be specified.");
                 exit(1);
             };
@@ -69,23 +80,96 @@ impl HandleCommand for InfoArgs {
         }
 
         // Show package version specific information
-        if let Some(package_id) = self.package.versioned() {
-            self.display_package_version_info(&package_id, &register, package);
+        if let Some(package_id) = package.versioned() {
+            self.display_package_version_info(&package_id, &register, installed_package);
             return;
         }
 
-        self.display_package_info(package);
+        self.display_package_info(&package.name, installed_package);
     }
 }
 
 impl InfoArgs {
+    /// Displays information about the Packit installation.
+    fn display_packit_info(&self, config: &Config, register: &PackageRegister) {
+        if self.tree {
+            error!(msg: "Displaying a tree requires package version to be specified.");
+            exit(1);
+        }
+
+        let target = Target::current();
+        let package_id = PackageId::new(PackageName::packit(), current_packit_version());
+
+        // Read latest version of packit
+        let manager = RepositoryManager::new(config);
+        let latest_version;
+        match manager.read_package(&package_id.name) {
+            Ok((repository_id, package_meta)) => match manager.read_latest_supported_version(&repository_id, &package_meta, &target) {
+                Ok(version_meta) => latest_version = Some(version_meta.version),
+                Err(e) => {
+                    error!(e, "Cannot request packit version metadata");
+                    latest_version = None;
+                },
+            },
+            Err(e) => {
+                error!(e, "Cannot request packit metadata");
+                latest_version = None;
+            },
+        }
+
+        // Create display for new version
+        let new_version_disp = match latest_version {
+            Some(latest_version) if latest_version > current_packit_version() => format!("yes, {}", latest_version.style()),
+            Some(_) | None => "no".into(),
+        };
+
+        println!("{}", package_id.style());
+        println!("{}", packit_version_name!().italic().cyan());
+
+        let mut pair_aligner = PairAligner::new();
+        pair_aligner.add("Config directory", DEFAULT_CONFIG_DIR);
+        pair_aligner.add("Prefix directory", config.prefix_directory.display());
+        pair_aligner.add("Multiuser mode", if config.multiuser { "on" } else { "off" });
+        pair_aligner.add("Installed packages", register.iterate_all().count());
+        pair_aligner.add("New version available", new_version_disp);
+        pair_aligner.display(PairAligner::VERTICAL_LINE_PREFIX);
+        println!();
+
+        println!("Current system");
+        let mut pair_aligner = PairAligner::new();
+
+        match &target.os {
+            OsVersion::MacOs { version } => {
+                pair_aligner.add("OS", format!("macOS {version}"));
+            },
+            OsVersion::Linux {
+                distro,
+                distro_version,
+                kernel_version,
+            } => {
+                pair_aligner.add("OS", format!("Linux {distro} {distro_version}"));
+                pair_aligner.add("Kernel version", kernel_version);
+            },
+            OsVersion::Windows { version } => {
+                pair_aligner.add("OS", format!("Windows {version}"));
+            },
+            OsVersion::Unknown => {
+                pair_aligner.add("OS", "Unknown".dimmed());
+            },
+        }
+
+        pair_aligner.add("Architecture", target.architecture);
+        pair_aligner.display(PairAligner::VERTICAL_LINE_PREFIX);
+        println!();
+    }
+
     /// Displays package info.
-    fn display_package_info(&self, package: &InstalledPackage) {
+    fn display_package_info(&self, package_name: &PackageName, package: &InstalledPackage) {
         // Sort installed versions for display
         let mut installed_versions: Vec<_> = package.versions.keys().collect();
         installed_versions.sort();
 
-        println!("{}", self.package.name.style());
+        println!("{}", package_name.style());
         println!("{}", package.description.italic().cyan());
 
         let mut pair_aligner = PairAligner::new();